+++ app/src/main/java/com/iic/mokojin/operations/LeaveQueueOperation.java
@@ -1,4 +1,4 @@
-package com.iic.mokojin.operation;
+package com.iic.mokojin.operations;
 
 import com.iic.mokojin.models.QueueItem;
 import com.parse.ParseCloud;
